fix/auth: fix again
