+++ src/screens/login/Login.tsx
@@ -1,10 +1,9 @@
 import React from 'react';
-import { View, Alert, Image, TouchableOpacity, Text } from 'react-native';
+import { View, Alert, Image } from 'react-native';
 import { Button } from '../../components/Button';
 import { TextInput } from '../../components/Form';
 import { useLogin } from '../../util/auth';
 import { styles } from './LoginStyles';
-import { login, logOut, register } from '../../util/auth2';
 
 const createTwoButtonAlert = () =>
   Alert.alert(
@@ -50,16 +49,6 @@ export const Login = () => {
         style={styles.img}
         source={require('../../../assets/profile.jpg')}
       />
-      <TouchableOpacity onPress={() => register('f4@jwlying.com', '123456')}>
-        <Text>register</Text>
-      </TouchableOpacity>
-      <TouchableOpacity onPress={() => login('tilman45@jwlying.com', '123456')}>
-        <Text>login</Text>
-      </TouchableOpacity>
-
-      <TouchableOpacity onPress={() => logOut()}>
-        <Text>log out</Text>
-      </TouchableOpacity>
 
       <TextInput
         label=""
